dan muresan warning patch

+++ vringbuffer.c
@@ -110,7 +110,7 @@ static void* my_malloc(size_t size1,size_t size2){
   {
     long pagesize = sysconf(_SC_PAGESIZE);
     char *cret=ret;
-    long i=0;
+    size_t i=0;
     for(i=0;i<size;i+=pagesize)
       cret[i]=0;
   }
